requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.7.0
pyscopg2==66.0.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
